fix:layout auth and public, layout auth tanpa footer, dan public memiliki footer
