+++ src/app/(auth)/layout.jsx
@@ -1,0 +1,22 @@
+import "../globals.css";
+import NavbarSection from "@/components/layout/Navbar";
+
+export const metadata = {
+  title: "Create Next App",
+  description: "Generated by create next app",
+};
+
+
+export default function AuthLayout({ children }) {
+  return (
+    <div className="antialiased flex flex-col min-h-screen">
+      <div className="bg-green-deep dark:bg-green-deep ">
+        <NavbarSection />
+      </div>
+
+      <section className="container max-w-9/10 mx-auto p-6 grow">
+        {children}
+      </section>
+    </div>
+  );
+}
