+++ src/app/(auth)/login/page.jsx
@@ -3,7 +3,7 @@ import React from 'react'
 
 const LoginPage = () => {
   return (
-    <div className='max-w-screen-xl mx-auto min-h-[30vh] md:min-h-[50vh] flex justify-center items-center py-10'>
+    <div className='max-w-screen-xl mx-auto min-h-[70vh] md:min-h-[80vh] flex justify-center items-center py-10'>
       <FormLogin/>
     </div>
   )
