+++ src/app/(auth)/register/page.jsx
