--- src/app/(home)/page.jsx
@@ -1,17 +1,0 @@
-
-import FormRegister from "@/components/layout/form/FormRegister";
-import { Button } from "@/components/ui/button";
-
-export default function Home() {
-
-
-  
-  return (
-    <div>
-      <h1>Landing Page</h1>
-      {/* <Button>Button aja</Button> */}
-      <FormRegister />
-      
-    </div>
-  );
-}
