+++ src/app/(public)/page.jsx
@@ -1,0 +1,9 @@
+
+
+export default function Home() {
+  return (
+    <div>
+      <h1>Landing Page</h1>
+    </div>
+  );
+}
