+++ src/app/(public)/snap/page.jsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+const SnapPage = () => {
+  return <div> SnapPage</div>;
+};
+
+export default SnapPage;
