+++ src/app/layout.jsx
@@ -1,28 +1,22 @@
 import "./globals.css";
 import FooterSection from "@/components/layout/Footer";
 import NavbarSection from "@/components/layout/Navbar";
-import { Urbanist } from 'next/font/google'
+import { Urbanist } from "next/font/google";
 
 export const metadata = {
   title: "Create Next App",
   description: "Generated by create next app",
 };
 
-const urbanist = Urbanist({ subsets: ["latin"], weight: ["100", "200", "300", "400", "500", "600", "700", "800", "900"] });
+const urbanist = Urbanist({
+  subsets: ["latin"],
+  weight: ["100", "200", "300", "400", "500", "600", "700", "800", "900"],
+});
 
 export default function RootLayout({ children }) {
   return (
     <html lang="en" className={urbanist.className}>
-      <body className="antialiased flex flex-col min-h-screen">
-        <div className="bg-green-deep dark:bg-green-deep ">
-          <NavbarSection />
-        </div>
-
-        <section className="container max-w-9/10 mx-auto p-6 grow">
-          {children}
-        </section>
-        <FooterSection />
-      </body>
+      <body>{children}</body>
     </html>
   );
 }
